Remove redundant PID and hostname code and move some code around.
